Simulate envelope detector
Create a seperate LTSpice file for the envelope detector. Design
includes an NPN transistor to replace the resistor that would allow
the capacitor to discharge. Thus, the discharge can be controlled
by a discharge pulse voltage, which in final design would likely be
implemented by the Arduino. Discharge for a pulse of 16.7 us (or
double the maximum inout frequency) can bring the voltage to 30 mV