Add firmware test1
Test the delay for the AD_9850.wr_serial() function to the propagation
of the new waveform.

This has important implications for the period for the discharge pin
as well as the setting of when to begin analogRead(). Take the AD9850
out and the TRIGGER pin as inputs to the oscilloscope, and measure the
difference between falling edge of the trigger and the change in the
output frequency.