Rearrange sweepStep interrupt function; Lint code
Move code to load array values to String buffer to before the frequency
is changed. Additionally, have the function return when 30000 Hz has
been met or exceeded. Lint unused functions. Code has yet to be tested
with new changes. May need to investigate whether every frequency step
is reached when printing, as it appears it may not be the case